util: rename util/unique to util/deduplicate
This commit renames the `util/unique` package to `util.deduplicate`.
This opens the way to add a new `unique` util package to handle unique
integer generation.

Epic: None

Release note: None
