+++ pkg/BUILD.bazel
@@ -699,6 +699,7 @@ ALL_TESTS = [
     "//pkg/util/ctxlog:ctxlog_test",
     "//pkg/util/ctxutil:ctxutil_test",
     "//pkg/util/debugutil:debugutil_test",
+    "//pkg/util/deduplicate:deduplicate_test",
     "//pkg/util/duration:duration_test",
     "//pkg/util/encoding/csv:csv_test",
     "//pkg/util/encoding:encoding_test",
@@ -777,7 +778,6 @@ ALL_TESTS = [
     "//pkg/util/tsearch:tsearch_test",
     "//pkg/util/uint128:uint128_test",
     "//pkg/util/ulid:ulid_test",
-    "//pkg/util/unique:unique_test",
     "//pkg/util/uuid:uuid_test",
     "//pkg/util/vector:vector_test",
     "//pkg/util/version:version_test",
@@ -2479,6 +2479,8 @@ GO_TARGETS = [
     "//pkg/util/ctxutil:ctxutil_test",
     "//pkg/util/debugutil:debugutil",
     "//pkg/util/debugutil:debugutil_test",
+    "//pkg/util/deduplicate:deduplicate",
+    "//pkg/util/deduplicate:deduplicate_test",
     "//pkg/util/duration:duration",
     "//pkg/util/duration:duration_test",
     "//pkg/util/encoding/csv:csv",
@@ -2668,8 +2670,6 @@ GO_TARGETS = [
     "//pkg/util/ulid:ulid",
     "//pkg/util/ulid:ulid_test",
     "//pkg/util/unaccent:unaccent",
-    "//pkg/util/unique:unique",
-    "//pkg/util/unique:unique_test",
     "//pkg/util/uuid:uuid",
     "//pkg/util/uuid:uuid_test",
     "//pkg/util/vector:vector",
