+++ pkg/sql/rowenc/BUILD.bazel
@@ -34,6 +34,7 @@ go_library(
         "//pkg/sql/sqlerrors",
         "//pkg/sql/types",
         "//pkg/util/buildutil",
+        "//pkg/util/deduplicate",
         "//pkg/util/encoding",
         "//pkg/util/intsets",
         "//pkg/util/json",
@@ -41,7 +42,6 @@ go_library(
         "//pkg/util/protoutil",
         "//pkg/util/trigram",
         "//pkg/util/tsearch",
-        "//pkg/util/unique",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_cockroachdb_redact//:redact",
     ],
