+++ pkg/util/json/BUILD.bazel
@@ -25,13 +25,13 @@ go_library(
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
         "//pkg/util/buildutil",
+        "//pkg/util/deduplicate",
         "//pkg/util/encoding",
         "//pkg/util/intsets",
         "//pkg/util/json/tokenizer",
         "//pkg/util/metamorphic",
         "//pkg/util/randutil",
         "//pkg/util/syncutil",
-        "//pkg/util/unique",
         "@com_github_cockroachdb_apd_v3//:apd",
         "@com_github_cockroachdb_errors//:errors",
     ],
@@ -50,10 +50,10 @@ go_test(
         "//pkg/sql/inverted",
         "//pkg/sql/pgwire/pgerror",
         "//pkg/testutils/datapathutils",
+        "//pkg/util/deduplicate",
         "//pkg/util/encoding",
         "//pkg/util/randutil",
         "//pkg/util/timeutil",
-        "//pkg/util/unique",
         "@com_github_cockroachdb_apd_v3//:apd",
         "@com_github_cockroachdb_redact//:redact",
         "@com_github_stretchr_testify//require",
