+++ pkg/util/json/json_test.go
@@ -18,10 +18,10 @@ import (
 	"github.com/cockroachdb/apd/v3"
 	"github.com/cockroachdb/cockroach/pkg/sql/inverted"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
+	"github.com/cockroachdb/cockroach/pkg/util/deduplicate"
 	"github.com/cockroachdb/cockroach/pkg/util/encoding"
 	"github.com/cockroachdb/cockroach/pkg/util/randutil"
 	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
-	"github.com/cockroachdb/cockroach/pkg/util/unique"
 	"github.com/cockroachdb/redact"
 	"github.com/stretchr/testify/require"
 )
@@ -1572,8 +1572,8 @@ func TestEncodeJSONInvertedIndex(t *testing.T) {
 		// Make sure that the expected encoding slice is sorted, as well as the
 		// output of the function under test, because the function under test can
 		// reorder the keys it's returning if there are arrays inside.
-		enc = unique.UniquifyByteSlices(enc)
-		c.expEnc = unique.UniquifyByteSlices(c.expEnc)
+		enc = deduplicate.ByteSlices(enc)
+		c.expEnc = deduplicate.ByteSlices(c.expEnc)
 		for j, path := range enc {
 			if !bytes.Equal(path, c.expEnc[j]) {
 				t.Errorf("unexpected encoding mismatch for %v. expected [%#v], got [%#v]",
